requirements.txt
# 核心依赖
neo4j>=5.0.0
psycopg2-binary>=2.9.0
pyodbc>=4.0.0
requests>=2.28.0
schedule>=1.2.0

# Web框架
Flask>=2.3.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 日志和配置
python-dotenv>=1.0.0

# 开发工具
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0